fixed modification of wrong engine configuration
