+++ modules/flowable5-engine/src/main/java/org/activiti/engine/impl/cfg/ProcessEngineConfigurationImpl.java
@@ -217,7 +217,6 @@ import org.flowable.image.impl.DefaultProcessDiagramGenerator;
 import org.flowable.job.service.impl.asyncexecutor.DefaultAsyncJobExecutor;
 import org.flowable.validation.ProcessValidator;
 import org.flowable.validation.ProcessValidatorFactory;
-import org.flowable.validation.validator.impl.ServiceTaskValidator;
 import org.flowable.variable.api.types.VariableType;
 import org.flowable.variable.api.types.VariableTypes;
 import org.flowable.variable.service.impl.types.BooleanType;
@@ -459,7 +458,6 @@ public abstract class ProcessEngineConfigurationImpl extends ProcessEngineConfig
     // PROCESS VALIDATION
 
     protected ProcessValidator processValidator;
-    protected ServiceTaskValidator customServiceTaskValidator;
 
     // OTHER ////////////////////////////////////////////////////////////////////
 
@@ -1504,13 +1502,7 @@ public abstract class ProcessEngineConfigurationImpl extends ProcessEngineConfig
 
     protected void initProcessValidator() {
         if (this.processValidator == null) {
-            if (customServiceTaskValidator == null) {
-                this.processValidator = new ProcessValidatorFactory().createDefaultProcessValidator();
-            } else {
-                ProcessValidatorFactory processValidatorFactory = new ProcessValidatorFactory();
-                processValidatorFactory.setCustomServiceTaskValidator(customServiceTaskValidator);
-                this.processValidator = processValidatorFactory.createDefaultProcessValidator();
-            }
+            this.processValidator = new ProcessValidatorFactory().createDefaultProcessValidator();
         }
     }
 
@@ -2208,14 +2200,6 @@ public abstract class ProcessEngineConfigurationImpl extends ProcessEngineConfig
 
     public void setProcessValidator(ProcessValidator processValidator) {
         this.processValidator = processValidator;
-    }
-
-    public ServiceTaskValidator getCustomServiceTaskValidator() {
-        return customServiceTaskValidator;
-    }
-
-    public void setCustomServiceTaskValidator(ServiceTaskValidator customServiceTaskValidator) {
-        this.customServiceTaskValidator = customServiceTaskValidator;
     }
     
     public XMLImporterFactory getWsdlImporterFactory() {
